Change enqueue scripts to enqueue scripts and styles
I really like this starter theme. However, the current GitHub version
is not enqueuing the stylesheets. In the version available on
WordPress.org (1.1.10), the stylesheets do enqueue. I noticed that in
version 1.1.10 scripts and styles are enqueued together. I made a local
copy that enqueues both at the same time and that seems to fix the
issue. Submitting these changes for your consideration.